Add function to render QR-Codes
Best practice of UX is not only to give user an error, but also a clear solution.
Showing a QR code when applicable is a good way to give a link to an updated
instructions, wiki, video, etc.

The added function allows rendering QR Code version 2 which can fit up to 47 alpha-
numeric characters, which is plenty if used with URL shortener.

This commit does not add any additional QR codes, only adds possiblity to do so.