Fix remaining time guestimation; Enable M73
Only attmept to calculate approximate remaining print time either after 1% of the file is done,
or at least after 1 minute of print time. Some longer prints will not advance past 1% even
in the first hour or so, which is not very user friendly.
Also, enable the M73 gcode parsing for this printer and use whatever values there is to show
more precise timings